feat: Refactor FeedScreen with new drag-and-drop mechanics and modular components for food management
